+++ gdk/win32/gdkdevice-wintab.c
@@ -46,10 +46,10 @@ static void gdk_device_wintab_query_state (GdkDevice        *device,
                                            GdkWindow        *window,
                                            GdkWindow       **root_window,
                                            GdkWindow       **child_window,
-                                           gint             *root_x,
-                                           gint             *root_y,
-                                           gint             *win_x,
-                                           gint             *win_y,
+                                           gdouble          *root_x,
+                                           gdouble          *root_y,
+                                           gdouble          *win_x,
+                                           gdouble          *win_y,
                                            GdkModifierType  *mask);
 static GdkGrabStatus gdk_device_wintab_grab   (GdkDevice     *device,
                                                GdkWindow     *window,
@@ -61,8 +61,8 @@ static GdkGrabStatus gdk_device_wintab_grab   (GdkDevice     *device,
 static void          gdk_device_wintab_ungrab (GdkDevice     *device,
                                                guint32        time_);
 static GdkWindow * gdk_device_wintab_window_at_position (GdkDevice       *device,
-                                                         gint            *win_x,
-                                                         gint            *win_y,
+                                                         gdouble         *win_x,
+                                                         gdouble         *win_y,
                                                          GdkModifierType *mask,
                                                          gboolean         get_toplevel);
 static void      gdk_device_wintab_select_window_events (GdkDevice       *device,
@@ -180,10 +180,10 @@ gdk_device_wintab_query_state (GdkDevice        *device,
                                GdkWindow        *window,
                                GdkWindow       **root_window,
                                GdkWindow       **child_window,
-                               gint             *root_x,
-                               gint             *root_y,
-                               gint             *win_x,
-                               gint             *win_y,
+                               gdouble          *root_x,
+                               gdouble          *root_y,
+                               gdouble          *win_x,
+                               gdouble          *win_y,
                                GdkModifierType  *mask)
 {
   GdkDeviceWintab *device_wintab;
@@ -268,8 +268,8 @@ gdk_device_wintab_ungrab (GdkDevice *device,
 
 static GdkWindow *
 gdk_device_wintab_window_at_position (GdkDevice       *device,
-                                      gint            *win_x,
-                                      gint            *win_y,
+                                      gdouble         *win_x,
+                                      gdouble         *win_y,
                                       GdkModifierType *mask,
                                       gboolean         get_toplevel)
 {
